+++ qiskit_aqt_provider/api_client/portal_client.py
@@ -10,14 +10,14 @@
 
 import os
 from collections.abc import Iterator
-from typing import Final, Optional
+from typing import Final, Optional, cast
 
 import httpx
+from aqt_connector.models.arnica.response_bodies.resources import ResourceDetails
 
 from qiskit_aqt_provider.api_client.errors import http_response_raise_for_status
 
 from . import models
-from . import models_generated as api_models
 from .versions import make_user_agent
 
 DEFAULT_PORTAL_URL: Final = httpx.URL("https://arnica.aqt.eu")
@@ -81,13 +81,13 @@ def _fetch_workspaces_summary(client: httpx.Client) -> models.ApiWorkspaces:
 def _fetch_resource(workspace_id: str, resource_id: str, client: httpx.Client) -> models.Resource:
     """Fetch all available information about a remote resource."""
     response = http_response_raise_for_status(client.get(f"/resources/{resource_id}"))
-    details = api_models.ResourceDetails.model_validate(response.json())
+    details = ResourceDetails.model_validate(response.json())
 
     return models.Resource(
         workspace_id=workspace_id,
         resource_id=resource_id,
         resource_name=details.name,
-        resource_type=details.type.value,
+        resource_type=cast(models.AQTBackendType, details.type.value),
         available_qubits=details.available_qubits,
     )